Create Networkx and network science
The following code is a python file which is an application of network science where algorithms such as page rank,betweenness centrality, degree centrality are calculated in order to understand the characteristics of the graph. In the end community detection algorithm is also implemented to classify the graph on the basis of modularity.

The following example is based on the dataset that I created which was based on the EU referrundum results that were conducted on 2016. A competition was cited among the different countries of UK where the country with maximum votes to leave the EU will have the major say in the final decision of UK. Hence, several games were cited and this code represents the game between England and Scotland. 

The data was collected from the official website of the electoral commission of UK, link = https://www.electoralcommission.org.uk/find-information-by-subject/elections-and-referendums/past-elections-and-referendums/eu-referendum/electorate-and-count-information